updated project showcase page styling
