+++ src/components/Pages/ProjectPortal/index.jsx
@@ -1,6 +1,6 @@
 import * as React from "react";
-import ProjectSideBar from './ProjectSideBar';
 import ProjectImages from './ProjectImages';
+import ProjectInfo from './ProjectInfo';
 
 class ProjectPortal extends React.Component {
   state = {
@@ -17,8 +17,7 @@ class ProjectPortal extends React.Component {
         <Banner />
         <Toolbar edit={this.edit} />
         <ProjectImages />
-        <ProjectDescription />
-        <ProjectSideBar />
+        <ProjectInfo />
         {this.state.isEditing && <EditProjectForm edit={this.edit} />}
       </div>
     );
@@ -51,53 +50,6 @@ function Toolbar({ edit }) {
     </div>
   );
 }
-
-
-function ProjectDescription() {
-  return (
-    <div className="project-portal__about">
-      <h2>Chingu</h2>
-      <p>August 16, 2018</p>
-      <div>
-        <p>
-          Chingu is building a global collaboration platform and coding-cohort
-          generator. A Chingu-cohort is a build-to-learn community, where
-          motivated developers from around the world, are organized into teams
-          to build a project in 8 weeks. We call these cohorts, Voyages. Chingu
-          so far has been a collaboration platform without an actual “platform”.
-          For each of the 50+ cohorts we’ve ran - which includes people from 135
-          different countries - we’ve managed to stay afloat using several
-          unconnected tools such as forms, spreadsheets, slack, github repos,
-          and trellos. With this many people, things tend to get a little messy
-          and chaotic -- for all of us. Our project for this competition aims to
-          fix this and to allow our thriving underground community to emerge.
-        </p>
-        <h3>Inspiration</h3>
-        <p>
-          Chingu is building a global collaboration platform and coding-cohort
-          generator. A Chingu-cohort is a build-to-learn community, where
-          motivated developers from around the world, are organized into teams
-          to build a project in 8 weeks. We call these cohorts, Voyages. Chingu
-          so far has been a collaboration platform without an actual “platform”.
-          For each of the 50+ cohorts we’ve ran - which includes people from 135
-          different countries - we’ve managed to stay afloat using several
-          unconnected tools such as forms, spreadsheets, slack, github repos,
-          and trellos. With this many people, things tend to get a little messy
-          and chaotic -- for all of us. Our project for this competition aims to
-          fix this and to allow our thriving underground community to emerge.
-        </p>
-        <h3>What we learned</h3>
-        <p>
-          That people are supercharged when optimally placed in an engaged
-          community with shared goals. And that this is especially true for
-          learners.
-        </p>
-      </div>
-    </div>
-  );
-}
-
-
 
 
 function EditProjectForm({ edit }) {
