+++ src/styles/pages/_project_portal.scss
@@ -3,12 +3,7 @@
 .project-portal {
   // TODO: Remove header positioning property
   padding-top: 64px + 21px; // to account for header
-  max-width: $max-width;
   margin: 0 auto;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-template-areas: "banner banner banner" "image image image-carousel" "description description cards";
-  grid-gap: 40px;
 }
 /*
 ==========================
@@ -18,13 +13,12 @@ BANNER WTIH PROJECT TITLE
 */
 .project-portal__banner {
   width: 100vw;
-  grid-area: banner;
-  transform:translateY(-21px); // to account for header
+  transform:translateY(-22px); // to account for header
   justify-self: center;
   text-align: center;
   color: #ffffff;
   min-height: 200px;
-  padding-top: 25px;
+  padding-top: 30px;
   background: linear-gradient(to right , #3b3c62,#74d6b1);
 }
 /*
@@ -50,64 +44,124 @@ TOOLBAR WITH PROJECT EDIT
 }
 /*
 ==========================
+OVERALL IMAGE CONTAINER
+==========================
+*/
+.project-image-container {
+  display: grid;
+  grid-template-areas: "image image image image-carousel image-carousel";
+  grid-template-columns: repeat(5, 1fr);
+  // max-height: 50vh;
+  height: 500px;
+}
+/*
+==========================
 IMAGE-CAROUSEL (PREVIEW)
 ==========================
 */
-.project-portal__preview {
+.project-portal__preview-color {
   grid-area: image-carousel;
+  background-color: $pale-grey;
+}
+.project-portal__preview {
+  width: 400px;
+  margin-left: 0px;
   display: grid;
   grid-gap: 5px;
   grid-template-columns: repeat(2, 1fr);
   grid-template-rows: auto;
-  max-height: 45vh;
-  overflow-x: hidden;
-  overflow-y: auto;
-  background-color: $pale-grey;
+  overflow-y: scroll;
+  padding: 30px;
+  height: 500px;
 }
-
 .project-portal__slide {
-  width: 85%;
+  width: 95%;
   margin: 0 auto;
 }
 
 .project-portal__img--preview {
   width: 100%;
-  height: 100px;
-  object-fit: contain;
+  height: 85px;
+  object-fit: cover;
   background-color: white;
   margin-top: 10px;
 }
 .project-portal__caption--preview {
   font-size: 12px;
   text-align: center;
-  color: $grey;
+  color: $dark-grey;
+  margin-top: -4px;
+}
+/*
+==========================
+UPLOAD IMAGE BTN
+==========================
+*/
+.project-upload-image-btn {
+  width: 0.1px;
+	height: 0.1px;
+	opacity: 0;
+	overflow: hidden;
+	position: absolute;
+	z-index: -1;
+}
+.project-upload-image-btn + label {
+  color: white;
+  background-color: $theme-green;
+  text-transform: uppercase;
+  margin: 0 auto;
+  grid-column: 1 / -1;
+  padding: 10px 30px;
+  margin-bottom: 20px;
+  border-radius: $border-radius;
+  cursor: pointer;
+  box-shadow: $box-shadow;
+  &:hover{
+    background-color: $theme-green--hover;
+  }
 }
 /*
 ==========================
 MAIN IMAGE
 ==========================
 */
-
-.project-portal__main-image {
+.project-portal__main-color {
   grid-area: image;
-  width: 100%;
   background-color: $palest-grey;
+  width: 100%;
+}
+.project-portal__main-image {
   margin: 0 auto;
+  width: 720px;
+  height: 100%;
+  margin-right: 10px;
 }
 .project-portal__slide--main {
   text-align: center;
-  margin: 0 auto;
+  height: 500px;
 }
 .project-portal__img--main {
   width: 650px;
-  height: 400px;
-  object-fit: contain;
+  height: 90%;
+  object-fit: cover;
 }
 .project-portal__caption--main {
-  margin-top: -50px;
+  color: $dark-grey;
+  margin-top: -4px;
 }
-
-
+/*
+==========================
+PROJECT INFO CONTAINER
+==========================
+*/
+.project-info-container {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-template-areas: "description description cards";
+  grid-gap: 40px;
+  max-width: $max-width;
+  margin: 0 auto;
+}
 /*
 ==========================
 CARDS (PROJECT SIDEBAR)
